% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressR.R
\name{compression}
\alias{compression}
\alias{compressZstd}
\alias{uncompressZstd}
\title{Compress and Decompress Objects Using the Zstandard algorithm}
\usage{
compressZstd(object, level = 3, raw = FALSE)

uncompressZstd(object, raw = FALSE)
}
\arguments{
\item{object}{Object to (un)compress.}

\item{level}{Integer. Compression level.}

\item{raw}{Logical. If FALSE, the \code{object} will be serialized before compression or after decompression.}
}
\value{
(un)compressed, and possibly (un)serialized, data.
}
\description{
Compress and decompress raw vectors and serialized objects using the Zstandard algorithm.
}
