% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressr-package.R
\docType{package}
\name{compressr-package}
\alias{compressr-package}
\alias{compressr}
\title{compressr package}
\description{
Simple wrapper functions for tar, untar, zip and unzip from the utils package.
}
\section{pack and unpack}{

The two primary functions used are \code{\link{pack}} and \code{\link{unpack}}. 
The former will package up the specified files or directory as either a 
\code{.tar.gz} or \code{.zip} file.  The latter will unpack a compressed file.
}

