% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{compressR}
\alias{compressR}
\title{Zstandard, or zstd, compression algorithm}
\description{
Zstandard, or zstd for short, is a fast lossless compression algorithm created by Facebook Inc.

  This package provides a simple interface to the Zstandard official C implementation.
}
