% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressr.R
\name{is_bzip2}
\alias{is_bzip2}
\title{Test if a file is gzipped}
\usage{
is_bzip2(filename)
}
\arguments{
\item{filename}{the fully qualified name of the file to test}
}
\value{
logical, TRUE if the input is likely gzipped
}
\description{
We guess that the first two bytes tell us enough to know it is bzip2
file we have.  For more information see \url{https://en.wikipedia.org/wiki/Bzip2#File_format}
and \url{https://www.ietf.org/rfc/rfc1952}
}
