% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressr.R
\name{is_gzip}
\alias{is_gzip}
\title{Test if a file is gzipped}
\usage{
is_gzip(filename)
}
\arguments{
\item{filename}{the fully qualified name of the file to test}
}
\value{
logical, TRUE if the input is likely gzipped
}
\description{
We guess that the first two bytes tell us enough to know what kind of zip
file we have.  For more information see 
\url{http://www.gzip.org/zlib/rfc-gzip.html#file-format}
and \url{https://www.ietf.org/rfc/rfc1952}
}
