% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressr.R
\name{is_zip}
\alias{is_zip}
\title{Test if a file is zipped.}
\usage{
is_zip(filename)
}
\arguments{
\item{filename}{the fully qualified name of the file to test}
}
\value{
logical, TRUE if the input is likely zipped
}
\description{
Test if a file is zipped.
}
