% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressr.R
\name{pack}
\alias{pack}
\title{Bundle and possibly compress a directory}
\usage{
pack(
  src,
  form = c("gzip", "zip")[1],
  filename = paste0(basename(src), ifelse(form[1] == "gzip", ".tar.gz", ".zip")),
  compression = "gzip",
  ...
)
}
\arguments{
\item{src}{character, path description for an exisiting directory}

\item{form}{character either 'gzip' or 'zip' to indicate compression used}

\item{filename}{character, path description for trhe destination file}

\item{compression}{character, see \code{\link[utils]{tar}} ignored if
\code{form} is not "gzip"}

\item{...}{arguments for \code{\link[utils]{tar}} or \code{\link[utils]{zip}}}
}
\value{
numeric with 0 for success as returned by \code{\link[utils]{tar}}
}
\description{
Bundle and possibly compress a directory
}
