% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save}
\alias{save}
\alias{saveZstd}
\alias{readZstd}
\title{Faster Serialization Interface for Single Objects}
\usage{
saveZstd(object, file, level = 3)

readZstd(file)
}
\arguments{
\item{object}{Object to compress.}

\item{file}{A connection or the name of the file where the \code{R} object is saved to or read from.}

\item{level}{Integer. Compression level.}
}
\value{
For \code{saveZstd}, NULL invisibly.

  For \code{readZstd}, the uncompressed and unserialized \code{R} object.
}
\description{
Functions to write a single \code{R} object to a file and
 to restore it using the Zstandard algorithm.
}
\seealso{
\code{\link{serialize}}, \code{\link{saveRDS}} and \code{\link{loadRDS}}.
}
