% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressr.R
\name{test_compressed}
\alias{test_compressed}
\title{Test if a given file is compressed by zip or tgz, gzip.}
\usage{
test_compressed(filename)
}
\arguments{
\item{filename}{the fully qualified name of the file to test}
}
\value{
a character vector of 'zip', 'gzip' or "tgz", "bzip2",  "unknown"
}
\description{
Test if a given file is compressed by zip or tgz, gzip.
}
\concept{COMPRESSED}
