% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressr.R
\name{unpack}
\alias{unpack}
\title{Uncompress a packed directory}
\usage{
unpack(filename, exdir = ".", ...)
}
\arguments{
\item{filename}{character, path description to the file to unpack}

\item{exdir}{character the path where to extract the package}

\item{...}{arguments for \code{\link[utils]{untar}} or \code{\link[utils]{unzip}}}
}
\value{
varies as returned from \code{\link[utils]{untar}} or 
        \code{\link[utils]{unzip}}
}
\description{
Uncompress a packed directory
}
